% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.lm.glm.jaqpot.R
\name{deploy.lm.glm.jaqpot}
\alias{deploy.lm.glm.jaqpot}
\title{deploy.on.jaqpot takes as input a trained model, it uploads it
after some prompt questions and returns a model id}
\usage{
deploy.lm.glm.jaqpot(object)
}
\description{
Suports lm from base
}
