% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.pbpk.jaqpot.R
\name{deploy.pbpk.jaqpot}
\alias{deploy.pbpk.jaqpot}
\title{deploy.pbpk.jaqpot takes as input the dataframe upon the predictions are made, in
order to create the dataset corresponding to the model}
\usage{
deploy.pbpk.jaqpot(dataframe, covariate_model, odes)
}
\description{
a covariate model wich calculates physiological parameters based
on a patient's characteristics
}
